+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+### 07/06/2018 - 1.1.0
+* Added new `Collection` methods, `sort()` and `sortBy()` (ability to sort the items in collection)
+* Added protected methods on `Collection` for `isCallable()` and `ValueCallable`.
+
 ### 06/07/2018 - 1.0.1
 * Minor updates to `Arr` class and added `Arr` tests for code better code coverage.
 * Changed the `collection->has()` to work with "dot" notation.